test(unit_tests): move pure integration tests to unit_tests/integration/
