+++ unit_tests/integration/test_kafka.py
@@ -18,7 +18,7 @@ from sdcm.stress_thread import CassandraStressThread
 from sdcm.kafka.kafka_cluster import LocalKafkaCluster
 from sdcm.kafka.kafka_consumer import KafkaCDCReaderThread
 from sdcm.utils.issues import SkipPerIssues
-from unit_tests.dummy_remote import LocalLoaderSetDummy
+from unit_tests.lib.dummy_remote import LocalLoaderSetDummy
 
 pytestmark = [
     pytest.mark.integration,
