+++ unit_tests/unit/nemesis/fake_cluster.py
@@ -4,7 +4,7 @@ from dataclasses import dataclass, field
 
 from sdcm.cluster import BaseScyllaCluster
 from sdcm.nemesis.utils.node_allocator import NemesisNodeAllocator
-from unit_tests.dummy_remote import LocalLoaderSetDummy
+from unit_tests.lib.dummy_remote import LocalLoaderSetDummy
 
 
 PARAMS = dict(nemesis_interval=1, nemesis_filter_seeds=False)
